refactor: modularize helpers and theme
