+++ utils/theme.py
@@ -1,0 +1,6 @@
+"""Theme configuration for the Dash application."""
+
+import dash_bootstrap_components as dbc
+
+# Central place to define the Bootstrap theme used across the app
+THEME = dbc.themes.DARKLY
